+++ API/paypal/paypal-client.js
@@ -13,7 +13,7 @@ const checkoutNodeJssdk = require('@paypal/checkout-server-sdk');
  * credentials have access.
  */
 function client(config) {
-    return new checkoutNodeJssdk.core.PayPalHttpClient(environment(config));
+  return new checkoutNodeJssdk.core.PayPalHttpClient(environment(config));
 }
 
 /**
@@ -23,39 +23,11 @@ function client(config) {
  *
  */
 function environment(config) {
-    const clientId = process.env.PAYPAL_CLIENT_ID || config.clientId;
-    const clientSecret = process.env.PAYPAL_CLIENT_SECRET || config.secret;
+  const clientId = process.env.PAYPAL_CLIENT_ID || config.clientId;
+  const clientSecret = process.env.PAYPAL_CLIENT_SECRET || config.secret;
 
-    const method = config.env ? `${capitalizeFirstLetter(config.env)}Environment` : 'SandboxEnvironment'
-    return new checkoutNodeJssdk.core[method](clientId, clientSecret);
+  const method = config.env ? `${capitalizeFirstLetter(config.env)}Environment` : 'SandboxEnvironment'
+  return new checkoutNodeJssdk.core[method](clientId, clientSecret);
 }
 
-function capitalizeFirstLetter(string) {
-    return string.charAt(0).toUpperCase() + string.slice(1);
-}
-
-async function prettyPrint(jsonData, pre="") {
-    let pretty = "";
-    function capitalize(string) {
-        return string.charAt(0).toUpperCase() + string.slice(1).toLowerCase();
-    }
-    for (let key in jsonData){
-        if (jsonData.hasOwnProperty(key)){
-            if (isNaN(key))
-              pretty += pre + capitalize(key) + ": ";
-            else
-              pretty += pre + (parseInt(key) + 1) + ": ";
-            if (typeof jsonData[key] === "object"){
-                pretty += "\n";
-                pretty += await prettyPrint(jsonData[key], pre + "    ");
-            }
-            else {
-                pretty += jsonData[key] + "\n";
-            }
-
-        }
-    }
-    return pretty;
-}
-
-module.exports = {client: client, prettyPrint: prettyPrint};
+module.exports = { client: client };