+++ types/PaypalState.ts
@@ -1,5 +1,6 @@
 // This object should represent structure of your modules Vuex state
 // It's a good practice is to name this interface accordingly to the KET (for example mailchimpState)
 export interface PaypalState {
-  trans: string[]
+  trans: string[],
+  methods: null
 }